% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couchDB.R
\name{couch_list_databases}
\alias{couch_list_databases}
\title{List available databases.}
\usage{
couch_list_databases(conn)
}
\arguments{
\item{conn}{A couchDB connection object}
}
\description{
Lists the available databases on the connection provided.
}

