% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couchDB.R
\name{couch_list_databases_url}
\alias{couch_list_databases_url}
\title{couch_list_databases_url}
\usage{
couch_list_databases_url(conn)
}
\arguments{
\item{conn}{A couchDB connection object.}
}
\description{
Format the url for fetching database-list for internal use
}

