% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couchDB.R
\name{print.couch_connection}
\alias{print.couch_connection}
\title{Print method for couchDB connection object}
\usage{
\method{print}{couch_connection}(conn)
}
\arguments{
\item{conn}{a couchDB connection object}
}
\description{
Prints the couchDB connection object.
}

