\name{couch_fetch_attachment}
\alias{couch_fetch_attachment}
\title{Fetch attachment}
\usage{
couch_fetch_attachment(conn, database, key, attachment)
}
\arguments{
  \item{conn}{A couchDB connection object}

  \item{database}{The database to connecto to}

  \item{key}{Key of document}

  \item{attachment}{name of attachment}
}
\description{
Gets a named attachment from a recors
}

