% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couchDB.R
\name{couch_fetch_attachment_url}
\alias{couch_fetch_attachment_url}
\title{Get attachment url}
\usage{
couch_fetch_attachment_url(conn, database, key = NULL, attachment = NULL)
}
\arguments{
\item{conn}{A connection object}

\item{database}{The database name}

\item{key}{Document key}

\item{attachment}{Name of the attachment}
}
\description{
Get the url for a specific attachment, This is sometimes useful
for direct reads to functions, in lieu of storing tempfiles.
}

